Fix rounding error

+++ src/main/java/training/busboard/Bus.java
@@ -1,7 +1,7 @@
 package training.busboard;
 
 public class Bus{
-    private double timeToStation;
+    private int timeToStation;
     private String id;
 
     public double getTimeToStation() {
@@ -17,7 +17,7 @@ public class Bus{
     }
 
     public Bus(int timeToStation, String id) {
-        this.timeToStation = timeToStation / 60.0;
+        this.timeToStation = timeToStation / 60;
         this.id = id;
     }
